demo(xychart): remove example changes for easier review
