+++ packages/vx-demo/src/sandboxes/vx-xychart/CustomChartBackground.tsx
@@ -6,6 +6,7 @@ const patternId = 'xy-chart-pattern';
 
 export default function CustomChartBackground() {
   const { theme, margin, width, height } = useContext(DataContext);
+  const textStyles = { ...theme?.axisStyles.x.bottom.axisLabel, textAnchor: 'start' };
 
   // early return values not available in context
   if (width == null || height == null || margin == null || theme == null) return null;
@@ -14,8 +15,8 @@ export default function CustomChartBackground() {
     <>
       <PatternLines
         id={patternId}
-        width={16}
-        height={16}
+        width={10}
+        height={10}
         orientation={['diagonal']}
         stroke={theme?.gridStyles?.stroke}
         strokeWidth={1}
@@ -27,8 +28,13 @@ export default function CustomChartBackground() {
         width={width - margin.left - margin.right}
         height={height - margin.top - margin.bottom}
         fill={`url(#${patternId})`}
-        fillOpacity={0.3}
       />
+      <text x={margin.left} y={height - margin.top + margin.bottom / 2} {...textStyles}>
+        width {width}px
+      </text>
+      <g transform={`translate(${margin.left / 1.3}, ${height - margin.bottom})rotate(-90)`}>
+        <text {...textStyles}>height {height}px</text>
+      </g>
     </>
   );
 }
