+++ packages/vx-demo/src/sandboxes/vx-xychart/Example.tsx
@@ -1,7 +1,8 @@
-import React from 'react';
+import React, { useState } from 'react';
 import cityTemperature, { CityTemperature } from '@vx/mock-data/lib/mocks/cityTemperature';
-import { AnimatedAxis, AnimatedGrid, DataProvider, LineSeries, XYChart } from '@vx/xychart';
-import ExampleControls from './ExampleControls';
+import { XYChart, LineSeries, DataProvider, darkTheme, XYChartTheme } from '@vx/xychart';
+
+import Controls from './Controls';
 import CustomChartBackground from './CustomChartBackground';
 
 type Props = {
@@ -9,56 +10,24 @@ type Props = {
   height: number;
 };
 
-const xScaleConfig = { type: 'time' } as const;
-const yScaleConfig = { type: 'linear' } as const;
-const numTicks = 4;
-const data = cityTemperature.slice(0, 100);
+const xScaleConfig = { type: 'linear' } as const;
+const yScaleConfig = xScaleConfig;
+const data = cityTemperature.slice(50, 80);
 const getDate = (d: CityTemperature) => new Date(d.date);
 const getSfTemperature = (d: CityTemperature) => Number(d['San Francisco']);
 
-export default function Example({ height }: Props) {
+export default function Example(_: Props) {
+  const [theme, setTheme] = useState<XYChartTheme>(darkTheme);
+
   return (
-    <ExampleControls>
-      {({
-        theme,
-        xAxisOrientation,
-        yAxisOrientation,
-        showGridRows,
-        showGridColumns,
-        animationTrajectory,
-      }) => (
-        <DataProvider theme={theme} xScale={xScaleConfig} yScale={yScaleConfig}>
-          <XYChart height={Math.min(400, height)}>
-            <CustomChartBackground />
-            <AnimatedAxis
-              key={`xaxis-${animationTrajectory}`} // force animate on update
-              orientation={xAxisOrientation}
-              numTicks={numTicks}
-              animationTrajectory={animationTrajectory}
-            />
-            <AnimatedAxis
-              key={`yaxis-${animationTrajectory}`}
-              label="Temperature (°F)"
-              orientation={yAxisOrientation}
-              numTicks={numTicks}
-              animationTrajectory={animationTrajectory}
-            />
-            <AnimatedGrid
-              key={`grid-${animationTrajectory}`}
-              rows={showGridRows}
-              columns={showGridColumns}
-              animationTrajectory={animationTrajectory}
-              numTicks={numTicks}
-            />
-            <LineSeries
-              dataKey="line"
-              data={data}
-              xAccessor={getDate}
-              yAccessor={getSfTemperature}
-            />
-          </XYChart>
-        </DataProvider>
-      )}
-    </ExampleControls>
+    <>
+      <DataProvider theme={theme} xScale={xScaleConfig} yScale={yScaleConfig}>
+        <XYChart height={400}>
+          <CustomChartBackground />
+          <LineSeries dataKey="line" data={data} xAccessor={getDate} yAccessor={getSfTemperature} />
+        </XYChart>
+      </DataProvider>
+      <Controls theme={theme} setTheme={setTheme} />
+    </>
   );
 }
